Add files via upload
Keyword extraction is to identify and recommend topical words and phrases from the content of documents. Keyword extraction applies in many areas especially when extracted keywords in the field of information retrieval (IR), is of particular interest because people retrieve helpful information according to keywords. However, a wide variety of keyword extraction methods are proposed and analyzed separately, the research of comparison of different keyword extraction methods is extremely needed.In our paper we have used Text rank algorithm to extract the keywords from a corpus.